requirements.txt
Django>=1.11.3
pytz>=2017.2
djangorestframework>=3.6.3
Pillow>=4.2.0
python-magic>=0.4.13
